Add quad support.
Only MemoryQuadDatasources will be able to provide quad patterns.

When the current datasource has quads, the QuadPatternRouter will be usable.

An N-Quads writer has been added to allow application/n-quads requests which extends the TurtleWriter.
JsonLdWriter now extends from NQuadsWriter since it is also able to write quads when the datasource allows this.
The HTMLWriter is implicitly changed to support quads as well.