Introduce views.
This is the first commit of a series to replace writers by views.
While a writer is responsible for many different types of responses
for a specific content type,
a view is responsible for only one type of responses
for a specific (family of) content types.
This makes the controller mechanism more extensible
towards new kinds of controllers and different content types.

This commit adds the base View classes
and implements NotFound as views instead of writers.