Controllers decide asynchronously whether to handle.
Previously, the handleRequest method would synchronously return
true or false to indicate whether it could handle a request or not.
This commit allows controllers to decide this asynchronously:
at the moment a controller sees it cannot handle a request,
it should invoke the next() callback.
If an error occurs, next() can be invoked with the error.