Add quad support to CompositeDatasource.
A "graph" option can be added to the settings
of each nested triples-only datasource,
which indicates the graph in which the triples exist.
If no "graph" option is added,
the triples will be available in the default graph.
If the nested datasource support quads,
no "graph" option should be added, its graphs will be inherited.